+++ CHANGELOG.md
@@ -48,6 +48,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Parse SV-specific frequencies key/values (#57)
 - Parse MEI-specific frequencies key/values (#58)
 - Parse cancer-specific `MSK_MVL` as `mvl_tag` (#59)
+- Parse Mivmir and Gicam scores for SNVs (#60)
 ### Changed
 - Structure of parsed entities, to reflect documents saved in the database (#20)
 - Improved and simplified code (#22)